Separate VanillaFix mixins into 'core' and 'optional'
'Core' mixins have "required: true" set in their Mixin config,
and are fully compatible with Sponge

'Optional' mixins do not have "required: true" set. Any mixins
in the 'mixins.optional.nosponge' package will be automatically skipped
when Sponge is installed, while mixins anywhere else in the
'mixins.optional' hierarchy will always be applied.

To make it easier to run Sponge and VanillaFix together in a development
environment, several @Shadow fields and methods are modified to have
'aliases'. This is due to the fact that Sponge and VanillaFix use
different versions of the MCP mappings. In production, this is not an
issue - but in development, one set of deobfuscated mappings will win.

Adding these aliases allows VanillaFix to run even when Sponge's version
of the MCP mappings 'wins' at development time. These aliases have no
effect in production, since the SRG name will always be checked first by
Mixin.